+++ tests/unit/test_git2sc.py
@@ -5,44 +5,47 @@ from git2sc.git2sc import Git2SC
 
 
 class TestGit2SC(unittest.TestCase):
+    '''Test class for the Git2SC class'''
 
     def setUp(self):
         self.api_url = 'https://confluence.sucks.com/wiki/rest/api'
         self.auth_string = 'user:password'
         self.auth = tuple(self.auth_string.split(':'))
-        self.g = Git2SC(self.api_url, self.auth_string)
+        self.git2sc = Git2SC(self.api_url, self.auth_string)
 
         self.requests_patch = patch('git2sc.git2sc.requests')
         self.requests = self.requests_patch.start()
 
-        self.print_patch = patch('git2sc.git2sc.print')
-        self.print = self.print_patch.start()
+        self.requests_error_patch = patch(
+            'git2sc.git2sc.Git2SC._requests_error',
+        )
+        self.requests_error = self.requests_error_patch.start()
 
     def tearDown(self):
-        self.requests.stop()
-        self.print.stop()
+        self.requests_patch.stop()
+        self.requests_error_patch.stop()
 
     def test_has_auth_set(self):
         'Required attribute for some methods'
 
-        self.assertEqual(self.g.auth, self.auth)
+        self.assertEqual(self.git2sc.auth, self.auth)
 
     def test_has_empty_pages_by_default(self):
         'Required to initialize the dictionary'
 
-        self.assertEqual(self.g.pages, {})
+        self.assertEqual(self.git2sc.pages, {})
 
     def test_has_confluence_url_set(self):
         'Required attribute for some methods'
 
-        self.assertEqual(self.g.api_url, self.api_url)
+        self.assertEqual(self.git2sc.api_url, self.api_url)
 
     def test_can_get_page_info(self):
         '''Required to ensure that the get_page_info method calls the correct
         api endpoint and returns a json object'''
 
         page_id = '372274410'
-        result = self.g.get_page_info(page_id)
+        result = self.git2sc.get_page_info(page_id)
         self.assertEqual(
             self.requests.get.assert_called_with(
                 '{}/content/{}?expand=ancestors,body.storage,version'.format(
@@ -53,7 +56,7 @@ class TestGit2SC(unittest.TestCase):
             ),
             None,
         )
-        self.assertTrue(self.requests.get.return_value.raise_for_status.called)
+        self.assertTrue(self.requests_error.called)
         self.assertEqual(result, self.requests.get.return_value.json())
 
     def test_can_get_space_homepage(self):
@@ -64,7 +67,7 @@ class TestGit2SC(unittest.TestCase):
         self.requests.get.return_value.json.return_value = {
             '_expandable': {'homepage': '/rest/api/content/372334010'},
         }
-        result = self.g.get_space_homepage(space_id)
+        result = self.git2sc.get_space_homepage(space_id)
         self.assertEqual(
             self.requests.get.assert_called_with(
                 '{}/space/{}'.format(
@@ -75,7 +78,7 @@ class TestGit2SC(unittest.TestCase):
             ),
             None,
         )
-        self.assertTrue(self.requests.get.return_value.raise_for_status.called)
+        self.assertTrue(self.requests_error.called)
         self.assertEqual(result, '372334010')
 
     def test_can_get_space_articles(self):
@@ -110,7 +113,7 @@ class TestGit2SC(unittest.TestCase):
                 "status": "current",
             },
         }
-        self.g.get_space_articles(space_id)
+        self.git2sc.get_space_articles(space_id)
         self.assertEqual(
             self.requests.get.assert_called_with(
                 '{}/content/?spaceKey={}?expand='
@@ -122,8 +125,8 @@ class TestGit2SC(unittest.TestCase):
             ),
             None,
         )
-        self.assertTrue(self.requests.get.return_value.raise_for_status.called)
-        self.assertEqual(self.g.pages, desired_pages)
+        self.assertTrue(self.requests_error.called)
+        self.assertEqual(self.git2sc.pages, desired_pages)
 
     def test_can_update_articles(self):
         '''Required to ensure that the update_page method posts to the
@@ -131,8 +134,8 @@ class TestGit2SC(unittest.TestCase):
 
         page_id = '372274410'
         html = '<p> This is a test </p>'
-        self.g.pages = {}
-        self.g.pages[page_id] = {
+        self.git2sc.pages = {}
+        self.git2sc.pages[page_id] = {
             'version': {
                 'number': 1
             },
@@ -146,7 +149,7 @@ class TestGit2SC(unittest.TestCase):
                 }
             ]
         }
-        self.g.update_page(page_id, html)
+        self.git2sc.update_page(page_id, html)
 
         data_json = json.dumps({
             'id': page_id,
@@ -175,7 +178,7 @@ class TestGit2SC(unittest.TestCase):
             ),
             None,
         )
-        self.assertTrue(self.requests.put.return_value.raise_for_status.called)
+        self.assertTrue(self.requests_error.called)
 
     @patch('git2sc.git2sc.json')
     @patch('git2sc.git2sc.Git2SC.get_page_info')
@@ -185,8 +188,8 @@ class TestGit2SC(unittest.TestCase):
 
         page_id = '372274410'
         html = '<p> This is a test </p>'
-        self.g.pages = {}
-        self.g.update_page(page_id, html)
+        self.git2sc.pages = {}
+        self.git2sc.update_page(page_id, html)
         self.assertEqual(getPageInfoMock.assert_called_with(page_id), None)
 
     def test_can_update_articles_with_title(self):
@@ -195,8 +198,8 @@ class TestGit2SC(unittest.TestCase):
 
         page_id = '372274410'
         html = '<p> This is a test </p>'
-        self.g.pages = {}
-        self.g.pages[page_id] = {
+        self.git2sc.pages = {}
+        self.git2sc.pages[page_id] = {
             'version': {
                 'number': 1
             },
@@ -210,8 +213,8 @@ class TestGit2SC(unittest.TestCase):
                 }
             ]
         }
-        self.g.update_page(page_id, html, 'new title')
-        self.assertEqual(self.g.pages[page_id]['title'], 'new title')
+        self.git2sc.update_page(page_id, html, 'new title')
+        self.assertEqual(self.git2sc.pages[page_id]['title'], 'new title')
 
     def test_can_create_articles_as_parent(self):
         '''Required to ensure that the create_page method posts to the
@@ -219,7 +222,7 @@ class TestGit2SC(unittest.TestCase):
         inheritance is set'''
 
         html = '<p> This is a new page </p>'
-        self.g.create_page('TST', 'new title', html)
+        self.git2sc.create_page('TST', 'new title', html)
 
         data_json = json.dumps({
             'type': 'page',
@@ -242,7 +245,7 @@ class TestGit2SC(unittest.TestCase):
             ),
             None,
         )
-        self.assertTrue(self.requests.post.return_value.raise_for_status.called)
+        self.assertTrue(self.requests_error.called)
 
     def test_can_create_articles_as_a_child(self):
         '''Required to ensure that the create_page method posts to the
@@ -251,7 +254,7 @@ class TestGit2SC(unittest.TestCase):
 
         html = '<p> This is a new page </p>'
         parent_id = '372274410'
-        self.g.create_page('TST', 'new title', html, parent_id)
+        self.git2sc.create_page('TST', 'new title', html, parent_id)
 
         data_json = json.dumps({
             'type': 'page',
@@ -276,5 +279,50 @@ class TestGit2SC(unittest.TestCase):
             ),
             None,
         )
-        self.assertTrue(self.requests.post.return_value.raise_for_status.called)
+        self.assertTrue(self.requests_error.called)
 
+
+class TestGit2SC_requests_error(unittest.TestCase):
+    '''Test class for the Git2SC _requests_error method'''
+
+    def setUp(self):
+        self.api_url = 'https://confluence.sucks.com/wiki/rest/api'
+        self.auth_string = 'user:password'
+        self.auth = tuple(self.auth_string.split(':'))
+        self.git2sc = Git2SC(self.api_url, self.auth_string)
+
+        self.requests_object = Mock()
+
+        self.print_patch = patch('git2sc.git2sc.print')
+        self.print = self.print_patch.start()
+
+    def tearDown(self):
+        self.print.stop()
+
+    def test_request_error_display_message_if_rc_not_200(self):
+        '''Required to ensure that the _requests_error method returns the
+        desired structure inside the print when a requests instance has a
+        return code different from 200'''
+
+        self.requests_object.text = json.dumps({
+            'statusCode': 400,
+            'message': 'Error message',
+        })
+        self.git2sc._requests_error(self.requests_object)
+        self.assertEqual(
+            self.print.assert_called_with(
+                'Error 400: Error message'
+            ),
+            None,
+        )
+
+    def test_request_error_do_nothing_if_rc_is_200(self):
+        '''Required to ensure that the _requests_error method does nothing
+        if the return code is 200'''
+
+        self.requests_object.text = json.dumps({
+            'statusCode': 200,
+            'message': 'Error message',
+        })
+        self.git2sc._requests_error(self.requests_object)
+        self.assertFalse(self.print.called)